scripts/requirements.txt
streamlit>=1.28.0
pandas>=1.5.0
numpy>=1.24.0
Pillow>=9.0.0
scikit-learn>=1.3.0
joblib>=1.3.0
opencv-python>=4.8.0
plotly>=5.15.0
requests>=2.31.0
